Refactorized idea controller
