+++ templates/base.html.twig
@@ -49,8 +49,8 @@
                 <div class="ui right dividing rail">
                     <div class="ui vertical stripe">
                         {% block side_menu %}
-                        {{ render(controller('App\\Controller\\IdeaController::getNextScheduledIdeas')) }}
-                        {{ render(controller('App\\Controller\\IdeaController::getMoreVotesPendingIdeas')) }}
+                        {{ render(controller('App\\Controller\\Idea\\NextScheduledIdeasEmbedController')) }}
+                        {{ render(controller('App\\Controller\\Idea\\MoreVotesPendingIdeasEmbedController')) }}
                         {% endblock %}
                     </div>
                 </div>
